Creation of WordPress Plugin
This update includes the initial import of the WordPress Plugin. The plugin picks out instances of the shortcode [music_library] and replaces it with the contents of the music library stored in the music library database.

The content displayed is dependent on the parameters passed in:

- If no parameters are passed in, all artists are displayed.
- If artist is passed in with no album, a list of all albums belonging to that artist are displayed
- If artist and album are passed in, the track listing for that album is displayed.

This is an initial version that requires much spit and polish.