Plugin - Now supports blogs running with and without permalinks.
Updated the code so it now works whether the user has permalinks enabled or not.

There is a known issue where if permalinks are disabled and you're viewing a subpage (ie. not the list of all artists), the urls generated on that page include duplicate fields. This however doesn't cause a failure.